refactor: migrate to feature-based project structure
