+++ server/src/main/java/com/express/inventory/api/logs/InventoryLogRepository.java
@@ -1,10 +1,10 @@
-package com.express.inventory.api.products;
+package com.express.inventory.api.logs;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.express.inventory.models.ProductEntity;
+import com.express.inventory.api.products.ProductEntity;
 
 public interface InventoryLogRepository extends JpaRepository<InventoryLogEntity, Integer> {
     // Get logs for a specific product
