+++ server/src/main/java/com/express/inventory/api/products/ProductRepository.java
@@ -1,10 +1,8 @@
-package com.express.inventory.repositories;
+package com.express.inventory.api.products;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.express.inventory.models.ProductEntity;
 
 public interface ProductRepository extends JpaRepository<ProductEntity, Integer> {
     public List<ProductEntity> findByNameContainingIgnoreCase(String name);
