+++ server/src/main/java/com/express/inventory/api/products/dto/response/ProductResponse.java
@@ -1,0 +1,5 @@
+package com.express.inventory.api.products.dto.response;
+
+public class ProductResponse {
+
+}
