+++ server/src/main/java/com/express/inventory/api/purchases/exception/InvalidPurchaseException.java
@@ -1,11 +1,11 @@
-package com.express.inventory.exceptions;
+package com.express.inventory.api.purchases.exception;
 
 public class InvalidPurchaseException extends RuntimeException {
     public InvalidPurchaseException(String message) {
         super(message);
     }
 
-        public InvalidPurchaseException() {
+    public InvalidPurchaseException() {
         super("Invalid purchase request. Please check the provided details and try again.");
     }
 }
