+++ server/src/main/java/com/express/inventory/api/purchases/exception/PurchaseNotFoundException.java
@@ -1,4 +1,4 @@
-package com.express.inventory.exceptions;
+package com.express.inventory.api.purchases.exception;
 
 public class PurchaseNotFoundException extends RuntimeException {
     public PurchaseNotFoundException(String message) {
