--- server/src/main/java/com/express/inventory/dto/products/response/ProductResponse.java
@@ -1,5 +1,0 @@
-package com.express.inventory.dto.products.response;
-
-public class ProductResponse {
-
-}
