+++ server/src/test/java/com/express/inventory/ExpressInventoryApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
-import com.express.inventory.services.ProductService;
+import com.express.inventory.api.products.ProductService;
 
 @SpringBootTest
 public class ExpressInventoryApplicationTests {
